pyproject.toml
[build-system]
requires = ["setuptools>=62.6",]
build-backend = "setuptools.build_meta"

[project]
name = "sunarp_service"
version = "2.4.0"
description = "It is a service that allows you to consult the data of the sunarp"
requires-python = ">=3.12"
authors = [{name="Leonardo Rojas"}]
readme = "README.md"
keywords = []
dependencies = [
    "patchright==1.56",
    "pydantic~=2.12",
    "loguru~=0.7",
    "python-stdnum==1.20",
    "pypdf==5.6.0",
]

[project.optional-dependencies]
test = [
"pytest==8.3.5",
"pytest-playwright==0.7.0",
]
dev = [
"mypy==1.15.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.*"]